+++ src/day12.rs
@@ -3,12 +3,11 @@ use crate::Solution;
 pub struct Day12;
 
 impl Solution for Day12 {
-
-    fn part1(input: &str) -> Option<usize> {
+    fn part1(input: &str) -> Option<i64> {
         None
     }
 
-    fn part2(input: &str) -> Option<usize> {
+    fn part2(input: &str) -> Option<i64> {
         None
     }
 }
