+++ src/day25.rs
@@ -3,12 +3,11 @@ use crate::Solution;
 pub struct Day25;
 
 impl Solution for Day25 {
-
-    fn part1(input: &str) -> Option<usize> {
+    fn part1(input: &str) -> Option<i64> {
         None
     }
 
-    fn part2(input: &str) -> Option<usize> {
+    fn part2(input: &str) -> Option<i64> {
         None
     }
 }
